back/requirements.txt
Flask==1.1.2
flask-mail==0.9.1
python-dotenv==1.0.0